+++ src/main/java/com/time/studentmanage/exception/TeacherExControllerAdvice.java
@@ -1,20 +1,26 @@
 package com.time.studentmanage.exception;
 
-import com.time.studentmanage.web.teacher.TeacherApiController;
-import lombok.extern.slf4j.Slf4j;
+import com.time.studentmanage.web.student.StudentController;
+import com.time.studentmanage.web.teacher.TeacherController;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@Slf4j
-@RestControllerAdvice(assignableTypes = TeacherApiController.class)
+@ControllerAdvice(assignableTypes = TeacherController.class)
 public class TeacherExControllerAdvice {
-    //IllegalArgumentException(500) -> BAD_REQUEST(400)
+
+    /**
+     * 접근 불가 페이지에 접근 시 처리하는 메서드
+     */
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ErrorResult IllegalExHandle(IllegalArgumentException e) {
-        log.error("[exceptionHandle] ex", e);
-        return new ErrorResult("BAD", e.getMessage());
+    @ExceptionHandler(value = {AccessDeniedException.class})
+    public String accessDeniedExceptionHandler(AccessDeniedException ex, Model model) {
+        model.addAttribute("errorMessage", ex.getMessage());
+        return "error/403";
     }
+
+
 }
